requirements.txt
aiohttp==3.6.0
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
schedule==0.6.0
slackclient==2.1.0
yarl==1.3.0
